Comparaison d'exercices, remplacement en séance, export/import fiabilisé
- Corrige l'export/import du commit précédent : poids décimaux préservés
  (57,5 kg ne devient plus 57), exercices réconciliés par nom (plus d'ids
  cassés), ordre des exercices conservé, vrai fichier .json partagé,
  dédoublonnage au ré-import, transaction unique
- Écran « Comparer deux exercices » : 2 courbes alignées sur les dates,
  accessible depuis Progression (bouton 📈)
- « Remplacer l'exercice » en séance : suggestions par muscles partagés,
  cibles et repos conservés
- Test aller-retour export/import (21 tests verts)

+++ lib/domain/services/export_service.dart
@@ -65,7 +65,7 @@ class ExportService {
                 'sets': [
                   for (final set in ex.sets)
                     {
-                      'weight': set.weightKg.toInt(),
+                      'weight': set.weightKg,
                       'reps': set.reps,
                       if ((set.durationSeconds ?? 0) > 0) 'durationSeconds': set.durationSeconds,
                       if ((set.distanceMeters ?? 0) > 0) 'distanceMeters': set.distanceMeters,
